Change some things
