+++ src/_default.page.server.tsx
@@ -3,7 +3,7 @@ import ReactDOMServer from 'react-dom/server';
 import React from 'react';
 import { PageShell } from './PageShell';
 import { escapeInject, dangerouslySkipEscape } from 'vite-plugin-ssr';
-import logoUrl from './logo.svg';
+import logoUrl from './assets/128x128.png';
 import type { PageContextServer } from './types';
 
 async function render(pageContext: PageContextServer) {
