+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import './index.css';
+import './App.scss';
 
 import { ChakraProvider, Theme, extendTheme } from '@chakra-ui/react';
